+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/models/Document.kt
@@ -1,0 +1,3 @@
+package org.vaccineimpact.orderlyweb.models
+
+data class Document(val displayName: String, val path: String, val isFile: Boolean, val children: List<Document>)
