+++ src/testHelpers/src/main/kotlin/org/vaccineimpact/orderlyweb/test_helpers/insertHelpers.kt
@@ -166,51 +166,3 @@ fun giveUserGroupGlobalPermission(db: JooqContext, userGroup: String, permission
         this.id = id
     }.insert()
 }
-
-fun insertDocuments()
-{
-    JooqContext().use {
-
-        it.dsl.insertInto(ORDERLYWEB_DOCUMENT)
-                .set(ORDERLYWEB_DOCUMENT.NAME, "some")
-                .set(ORDERLYWEB_DOCUMENT.PATH, "/some/")
-                .set(ORDERLYWEB_DOCUMENT.IS_FILE, 0)
-                .execute()
-
-        it.dsl.insertInto(ORDERLYWEB_DOCUMENT)
-                .set(ORDERLYWEB_DOCUMENT.NAME, "root")
-                .set(ORDERLYWEB_DOCUMENT.PATH, "/root/")
-                .set(ORDERLYWEB_DOCUMENT.IS_FILE, 0)
-                .execute()
-
-        it.dsl.newRecord(ORDERLYWEB_DOCUMENT)
-                .apply {
-                    this.name = "path"
-                    this.path = "/some/path/"
-                    this.parent = "/some/"
-                    this.isFile = 0
-                }.insert()
-
-        it.dsl.newRecord(ORDERLYWEB_DOCUMENT)
-                .apply {
-                    this.name = "file.csv"
-                    this.path = "/some/path/file.csv"
-                    this.parent = "/some/path/"
-                }.insert()
-
-        it.dsl.newRecord(ORDERLYWEB_DOCUMENT)
-                .apply {
-                    this.name = "first.csv"
-                    this.path = "/some/first.csv"
-                    this.parent = "/some/"
-                }.insert()
-
-        it.dsl.newRecord(ORDERLYWEB_DOCUMENT)
-                .apply {
-                    this.name = "empty"
-                    this.path = "/some/empty/"
-                    this.parent = "/some/"
-                    this.isFile = 0
-                }.insert()
-    }
-}
